Move files to regridding module
